requirements.txt
streamlit==1.28.1
openai==1.3.7
python-dotenv==1.0.0
langgraph==0.6.7
langchain==0.3.27
langchain-openai==0.3.33
langchain-core==0.3.27
